requirements.txt
torch
torchvision
transformers
accelerate
peft
pillow
pyyaml
imageio
imageio-ffmpeg
tqdm